requirements.dev.txt
black==22.3.0
djangorestframework==3.13.1
drf-yasg==1.20.0
ipython==8.3.0
isort==5.10.1
pyflakes==2.4.0
pylama==8.3.8